refactored flux calculations into a function
I replaced the repeated blocks of code that calculate the emission line fluxes, errors, and equivalent widths with a 'calculate_emission_line_flux()' function. There are still special cases for doublets and blended lines that have not been refactored. I also refactored repeated lines of code that mask emission lines when making a first estimation of the continuum fit, and replaced them with a 'mask_emission_lines()' function. The extraneous blocks of code have been commented out for removal in the next commit, and previously commented out blocks from the last commit have been removed completely. I also updated the conversion between FWHM and sigma from a factor of 2.35 to 2.3548 for higher precision.