major refactoring into functions for repeat tasks
The code has been completely refactored to extract several functions that more efficiently accomplish repeated tasks. These include estimating the emission line amplitudes and calculating the emission line fluxes. In addition, all of the model parameter indices are now defined at the top of the code, so there are no hard-coded model parameters during execution. This will allow additional emission lines to be added and/or removed more easily without having to make major changes to the main code. Finally, several of the parameter limits have been modified for use with the MUSE data, e.g. the [N II]/Ha ratio is not fixed when the lines are present in MUSE, because they are fully resolved and can be fit independently. Similarly, the density dependent [S II] and [O II] line ratios are now allowed to vary between their low and high density limits. Several tweaks may be required after testing on a larger number of objects.