added uv emission lines, fast fit mode, and refactored
The emission line model was updated to include a total of 34 emission lines, including 17 new ultraviolet and optical lines. The changes include all steps to define the rest wavelengths and model parameters, mask the lines during continuum fitting, calculate their redshifted positions, and save the flux and error results. The physically plausible ratio ranges for the new doublet lines were determined from grids of cloudy models. A half-gaussian was added to fit the asymmetric wing of lyman alpha. The centroid is locked to the primary component, with the height allowed to vary. The initial guesses may be uncertain.

The large number of emission lines slows the fitting process, and a 'fast_fit' mode was added for testing. This fixes the relative dispersion ratio, the gaussian height values to the initial guesses, and fixes all 'dz' values to zero. The full fit model is currently the default, and the modes can be accessed by typing 'full' and 'fast' in the terminal during fitting. By default, the code will warn the user if they attempt to save the results of a fast fit to the catalog. A timing statement was added to the code to assess performance.

Several improvements and corrections have been implemented. These include correcting the spectral dispersion ratio (MUSE/G141) from 0.027 to 0.058, which is required because the grism spectra are over-sampled by a factor of ~2 during the extraction by Nor Pirzkal. In addition, this value is allowed to vary lower by up to a factor of two, in order to account for broadening of the grism lines when objects are spatially extended. Extensive testing showed that this is sufficient, even for very large objects.

The size of each object is read into the code and passed to the fitting routine by the 'ab_image_max' variable, but testing showed that the fit successfully converges with the default initial dispersion ratio guess, and so has not been implemented further. In addition, the continuum fitting model was modified to allow for a linear fit to the grism data, which is sometimes more appropriate than the higher-order spline fitting model.

The code was refactored so that the initial values of some model parameters are set using a single loop rather than multiple hard-coded lines. For example, the 'dz' redshift offset guesses are all initially zero. Lastly, an amplitude index error was corrected in the 'calculate_doublet_line_flux()' function, and the fluxes and errors for each line saved in 'fit_results' are calculated by looping over a list of 'result_lines', rather than using many hard-coded lines.