added new emission lines, zoom-in plots, and refactored
The code has been updated to include a total of 34 emission lines, and add zoom-in plots for emission lines in MUSE that are fit with a S/N > 3. This helps to visually assess the quality of the fit for the high dispersion MUSE data. In addition, all of the line specific variables, such as lists of the line names and strings, have been moved to the top of the code for easier maintaining.

The default guess line is now [O II] 3727, which is commonly seen in MUSE for many targets. The 'fast_fit' mode has been added to the 'fit_inputs', as described in the last commit for 'fitting.py'. Finally, the plot styling has been updated to optimize the white space, axes labels, filenames, and zoom-in plots, and the results are now written to the catalog using a loop, saving ~200 lines of code.