added 'dz' option to change 'delta_z' interactively
The user can now type 'dz' while fitting a spectrum to change the 'delta_z' value that is initially set in the default.config file. This is useful for allowing a larger redshift range between widely separated emission lines in a small number of spectra. The value of 'delta_z' will reset to that in the default.config file after the object has been completed.

In addition, a small bug was fixed where the code would crash if > 7 emission lines were detected in MUSE with S/N > 3, which arose when trying to make the zoom-in plots.