commented out unused lines
The database manager system is now mostly deprecated, and the outdated sections have been commented out to prevent possible run issues. In the future, this should be updated to match the format of 'measure_z_interactive.py', or removed completely if the dependence is no longer required by the code to properly save the catalog results.