updated equation for zguess
The equation for 'zguess' was updated so the code centers the redshift guess on each of the [O II] 3727, 3730 doublet lines, instead of their average wavelength value. In addition, the 'fast' and 'full' options were moved inside the main fitting loop so an invalid option message is not printed when typing these commands. Finally, the format commands for printing the redshift error to the catalog were updated, but there is still a bug where no space is printed between the redshift error and right ascension values.