updated format commands for writing to catalog
This update fixed a bug where some columns were written together as a single number in the catalog, which prevented the values from being parsed. Specifically, the redshift error and right ascension were occasionally merged into a single number, as well as the 'snr_tot_others' and 'chisq' values. In addition, the redshift and redshift uncertainty are now saved to five decimal places for higher precision. Finally, a bug was fixed that occasionally led to infinities for the value of 'snr_tot_others'.